requirements.txt
pygame>=2.5
neat-python>=0.92
numpy>=1.24
